requirements.txt
# =========================================
# COLISSO - MODULE 1 : CORE
# =========================================
# Packages essentiels seulement

# Django
Django==5.0.0

# Django REST Framework
djangorestframework==3.14.0

# API Tools
django-filter==23.5
drf-spectacular==0.27.0

# CORS
django-cors-headers==4.3.1

# Configuration
python-decouple==3.8
